+++ Coral.Native/Source/Coral/PropertyInfo.cpp
@@ -1,8 +1,9 @@
-#include "PropertyInfo.hpp"
-#include "Type.hpp"
-#include "Attribute.hpp"
+#include "Coral/PropertyInfo.hpp"
+#include "Coral/Type.hpp"
+#include "Coral/Attribute.hpp"
+#include "Coral/TypeCache.hpp"
+
 #include "CoralManagedFunctions.hpp"
-#include "TypeCache.hpp"
 
 namespace Coral {
 
@@ -27,7 +28,8 @@ namespace Coral {
 	{
 		int32_t attributeCount;
 		s_ManagedFunctions.GetPropertyInfoAttributesFptr(m_Handle, nullptr, &attributeCount);
-		std::vector<ManagedHandle> attributeHandles(attributeCount);
+
+		std::vector<ManagedHandle> attributeHandles(static_cast<size_t>(attributeCount));
 		s_ManagedFunctions.GetPropertyInfoAttributesFptr(m_Handle, attributeHandles.data(), &attributeCount);
 
 		std::vector<Attribute> result(attributeHandles.size());
